+++ nengo_dl/tests/test_learning_rules.py
@@ -1,7 +1,11 @@
+from functools import partial
+
 import nengo
 from nengo.builder.learning_rules import SimVoja, SimOja, SimBCM
 import numpy as np
 import pytest
+
+from nengo_dl import configure_settings, graph_optimizer
 
 
 @pytest.mark.parametrize("rule", (nengo.Voja, nengo.Oja, nengo.BCM))
@@ -10,10 +14,13 @@ def test_merged_learning(Simulator, rule, seed):
     # make sure that works OK
     dimensions = 2
     with nengo.Network(seed=seed) as net:
-        a = nengo.Ensemble(3, dimensions)
-        b = nengo.Ensemble(3, dimensions)
-        c = nengo.Ensemble(5, dimensions)
-        d = nengo.Ensemble(10, dimensions)
+        configure_settings(
+            planner=partial(graph_optimizer.tree_planner, max_depth=10))
+
+        a = nengo.Ensemble(3, dimensions, label="a")
+        b = nengo.Ensemble(3, dimensions, label="b")
+        c = nengo.Ensemble(5, dimensions, label="c")
+        d = nengo.Ensemble(10, dimensions, label="d")
 
         conn0 = nengo.Connection(
             a, c, learning_rule_type=rule(),