+++ nengo_dl/utils.py
@@ -8,8 +8,8 @@ import time
 import warnings
 
 from nengo import Connection, Ensemble, Network, ensemble
-from nengo.exceptions import SimulationError, ConfigError
-from nengo.params import BoolParam
+from nengo.exceptions import SimulationError, ConfigError, NetworkContextError
+from nengo.params import BoolParam, Parameter
 import numpy as np
 import tensorflow as tf
 from tensorflow.python.framework.ops import get_gradient_function
@@ -30,7 +30,7 @@ else:
 
 
 def sanitize_name(name):
-    """Remove illegal Tensorflow name characters from string.
+    """Remove illegal TensorFlow name characters from string.
 
     Valid Tensorflow name characters are ``[A-Za-z0-9_.\\-/]``
 
@@ -64,7 +64,7 @@ def function_name(func, sanitize=True):
     func : callable
         callable object (e.g., function, callable class)
     sanitize : bool, optional
-        if True, remove any illegal Tensorflow name characters from name
+        if True, remove any illegal TensorFlow name characters from name
 
     Returns
     -------
@@ -83,9 +83,6 @@ def align_func(output_shape, output_dtype):
     """Decorator that ensures the output of ``func`` is an
     :class:`~numpy:numpy.ndarray` with the given shape and dtype.
 
-    Raises a ``SimulationError`` if the function returns ``None`` or a
-    non-finite value.
-
     Parameters
     ----------
     output_shape : tuple of int
@@ -93,6 +90,11 @@ def align_func(output_shape, output_dtype):
         function output)
     output_dtype : ``tf.DType`` or :class:`~numpy:numpy.dtype`
         desired dtype of function output
+
+    Raises
+    ------
+    :class:`~nengo:nengo.exceptions.SimulationError`
+        If the function returns ``None`` or a non-finite value.
     """
 
     if isinstance(output_dtype, tf.DType):
@@ -120,7 +122,7 @@ def align_func(output_shape, output_dtype):
 
 
 def print_op(input, message):
-    """Inserts a print statement into the tensorflow graph.
+    """Inserts a print statement into the TensorFlow graph.
 
     Parameters
     ----------
@@ -156,7 +158,7 @@ def cast_dtype(dtype, target):
     """Changes float dtypes to the target dtype, leaves others unchanged.
 
     Used to map all float values to a target precision.  Also casts numpy
-    dtypes to Tensorflow dtypes.
+    dtypes to TensorFlow dtypes.
 
     Parameters
     ----------
@@ -181,7 +183,7 @@ def cast_dtype(dtype, target):
 
 
 def find_non_differentiable(inputs, outputs):
-    """Searches through a Tensorflow graph to find non-differentiable elements
+    """Searches through a TensorFlow graph to find non-differentiable elements
     between ``inputs`` and ``outputs`` (elements that would prevent us from
     computing ``d_outputs / d_inputs``.
 
@@ -344,33 +346,58 @@ def minibatch_generator(inputs, targets, minibatch_size, shuffle=True,
                {p: targets[p][perm[i:i + minibatch_size]] for p in targets})
 
 
-def configure_trainable(config, default=None):
-    """Adds a configurable attribute called ``trainable`` to trainable objects.
+def configure_settings(**kwargs):
+    """Pass settings to ``nengo_dl`` by setting them as parameters on the
+    top-level Network config.
 
-    Used to manually configure whether or not those parts of the model can
-    be optimized by :meth:`.Simulator.train`.
+    The settings are passed as keyword arguments to ``configure_settings``;
+    e.g., to set ``trainable`` use ``configure_settings(trainable=True)``.
 
     Parameters
     ----------
-    config : :class:`~nengo:nengo.config.Config` or \
-             :class:`~nengo:nengo.Network`
-        the config object to be modified (or a Network to modify
-        ``net.config``)
-    default : bool, optional
-        the default value for ``trainable`` (``None`` means that the value
-        is deferred to the parent config, or ``True`` if this is the top-level
-        config)
+    trainable : bool or None
+        Adds a parameter to Nengo Ensembles/Connections/Networks that controls
+        whether or not they will be optimized by :meth:`.Simulator.train``.
+        Passing ``None`` will use the default ``nengo_dl`` trainable settings,
+        or True/False will override the default for all objects.  In either
+        case trainability can be further configured on a per-object basis (e.g.
+        ``net.config[my_ensemble].trainable = True``.  See `the documentation
+        <https://nengo.github.io/nengo_dl/training.html#choosing-which-elements-to-optimize>`_
+        for more details.
+    planner : graph planning algorithm
+        Pass one of the `graph planners
+        <https://nengo.github.io/nengo_dl/graph_optimizer.html>`_ to change the
+        default planner.
     """
 
-    if isinstance(config, Network):
-        config = config.config
+    # get the toplevel network
+    if len(Network.context) > 0:
+        config = Network.context[0].config
+    else:
+        raise NetworkContextError(
+            "``configure_settings`` must be called within a Network context "
+            "(``with nengo.Network(): ...``)")
 
-    for obj in (Ensemble, Connection, ensemble.Neurons):
-        try:
-            params = config[obj]
-        except ConfigError:
-            config.configures(obj)
-            params = config[obj]
+    try:
+        params = config[Network]
+    except ConfigError:
+        config.configures(Network)
+        params = config[Network]
 
-        params.set_param("trainable", BoolParam("trainable", default,
-                                                optional=True))
+    for attr, val in kwargs.items():
+        # TODO: we could prefix attr with "nengo_dl" or something if we're
+        # worried about conflicts (but since Networks aren't even configurable
+        # by default I'm not too concerned)
+
+        if attr == "trainable":
+            for obj in (Ensemble, Connection, ensemble.Neurons, Network):
+                try:
+                    obj_params = config[obj]
+                except ConfigError:
+                    config.configures(obj)
+                    obj_params = config[obj]
+
+                obj_params.set_param("trainable", BoolParam("trainable", val,
+                                                            optional=True))
+        else:
+            params.set_param(attr, Parameter(attr, val))